// src/types/bindings/CompanyProfileRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CompanyProfileRow = { id: bigint, company_name: string | null, legal_name: string | null, gstin: string | null, pan: string | null, address1: string | null, address2: string | null, location: string | null, pincode: string | null, state_code: string | null, phone: string | null, email: string | null, logo: string | null, };